appearance: update to new plugin format
